Undo Redo: Improve performance when undo / redo stack changes
I noticed while profiling my code that 99% of the CPU time to change the
undo redo stack is spent in Gemini instead of actually setting values.

This commit changes the undo / redo stack into a single stack with an
integer keeping track of what is undo and what is redo. This way we can
just easily synchronize the HistoryViewModel from the CollectionChanged
event of the UndoRedoManager stack. The advantage is that we can use an
item index in HistoryViewModel that is derived from the UndoRedoManager's
item index.

To do an undo / redo, we only need do undo / execute the needed actions
and update the UndoActions. The HistoryViewModel listens for UndoActions
changes and will update all the HistoryItemViewModel's ItemTypes. This has
the advantage of only refreshing the view once when all the actions have
been completed.

This greatly improves performance for large undo / redo operations.